move checkAdUnitSetup and refactor video.js hook
