+++ test/spec/unit/core/adapterManager_spec.js
@@ -1065,210 +1065,4 @@ describe('adapterManager tests', function () {
       });
     });
   });
-
-  describe('isValidBidRequest', function () {
-    describe('positive tests for validating bid request', function () {
-      beforeEach(function () {
-        sinon.stub(utils, 'logInfo');
-      });
-
-      afterEach(function () {
-        utils.logInfo.restore();
-      });
-
-      it('should maintain adUnit structure and adUnits.sizes is replaced', function () {
-        let fullAdUnit = [{
-          sizes: [[300, 250], [300, 600]],
-          mediaTypes: {
-            banner: {
-              sizes: [[300, 250]]
-            },
-            video: {
-              playerSize: [[640, 480]]
-            },
-            native: {
-              image: {
-                sizes: [150, 150],
-                aspect_ratios: [140, 140]
-              },
-              icon: {
-                sizes: [75, 75]
-              }
-            }
-          }
-        }];
-        let result = checkAdUnitSetup(fullAdUnit);
-        expect(result[0].sizes).to.deep.equal([[640, 480]]);
-        expect(result[0].mediaTypes.video.playerSize).to.deep.equal([[640, 480]]);
-        expect(result[0].mediaTypes.native.image.sizes).to.deep.equal([150, 150]);
-        expect(result[0].mediaTypes.native.icon.sizes).to.deep.equal([75, 75]);
-        expect(result[0].mediaTypes.native.image.aspect_ratios).to.deep.equal([140, 140]);
-
-        let noOptnlFieldAdUnit = [{
-          sizes: [[300, 250], [300, 600]],
-          mediaTypes: {
-            banner: {
-              sizes: [[300, 250]]
-            },
-            video: {
-              context: 'outstream'
-            },
-            native: {
-              image: {
-                required: true
-              },
-              icon: {
-                required: true
-              }
-            }
-          }
-        }];
-        result = checkAdUnitSetup(noOptnlFieldAdUnit);
-        expect(result[0].sizes).to.deep.equal([[300, 250]]);
-        expect(result[0].mediaTypes.video).to.exist;
-
-        let mixedAdUnit = [{
-          sizes: [[300, 250], [300, 600]],
-          mediaTypes: {
-            video: {
-              context: 'outstream',
-              playerSize: [[400, 350]]
-            },
-            native: {
-              image: {
-                aspect_ratios: [200, 150],
-                required: true
-              }
-            }
-          }
-        }];
-        result = checkAdUnitSetup(mixedAdUnit);
-        expect(result[0].sizes).to.deep.equal([[400, 350]]);
-        expect(result[0].mediaTypes.video).to.exist;
-
-        let altVideoPlayerSize = [{
-          sizes: [[600, 600]],
-          mediaTypes: {
-            video: {
-              playerSize: [640, 480]
-            }
-          }
-        }];
-        result = checkAdUnitSetup(altVideoPlayerSize);
-        expect(result[0].sizes).to.deep.equal([[640, 480]]);
-        expect(result[0].mediaTypes.video.playerSize).to.deep.equal([[640, 480]]);
-        expect(result[0].mediaTypes.video).to.exist;
-        sinon.assert.calledOnce(utils.logInfo);
-      });
-
-      it('should normalize adUnit.sizes and adUnit.mediaTypes.banner.sizes', function () {
-        let fullAdUnit = [{
-          sizes: [300, 250],
-          mediaTypes: {
-            banner: {
-              sizes: [300, 250]
-            }
-          }
-        }];
-        let result = checkAdUnitSetup(fullAdUnit);
-        expect(result[0].sizes).to.deep.equal([[300, 250]]);
-        expect(result[0].mediaTypes.banner.sizes).to.deep.equal([[300, 250]]);
-      });
-    });
-
-    describe('negative tests for validating bid requests', function () {
-      beforeEach(function () {
-        sinon.stub(utils, 'logError');
-      });
-
-      afterEach(function () {
-        utils.logError.restore();
-      });
-
-      it('should throw error message and delete an object/property', function () {
-        let badBanner = [{
-          sizes: [[300, 250], [300, 600]],
-          mediaTypes: {
-            banner: {
-              name: 'test'
-            }
-          }
-        }];
-        let result = checkAdUnitSetup(badBanner);
-        expect(result[0].sizes).to.deep.equal([[300, 250], [300, 600]]);
-        expect(result[0].mediaTypes.banner).to.be.undefined;
-        sinon.assert.called(utils.logError);
-
-        let badVideo1 = [{
-          sizes: [[600, 600]],
-          mediaTypes: {
-            video: {
-              playerSize: ['600x400']
-            }
-          }
-        }];
-        result = checkAdUnitSetup(badVideo1);
-        expect(result[0].sizes).to.deep.equal([[600, 600]]);
-        expect(result[0].mediaTypes.video.playerSize).to.be.undefined;
-        expect(result[0].mediaTypes.video).to.exist;
-        sinon.assert.called(utils.logError);
-
-        let badVideo2 = [{
-          sizes: [[600, 600]],
-          mediaTypes: {
-            video: {
-              playerSize: [['300', '200']]
-            }
-          }
-        }];
-        result = checkAdUnitSetup(badVideo2);
-        expect(result[0].sizes).to.deep.equal([[600, 600]]);
-        expect(result[0].mediaTypes.video.playerSize).to.be.undefined;
-        expect(result[0].mediaTypes.video).to.exist;
-        sinon.assert.called(utils.logError);
-
-        let badNativeImgSize = [{
-          mediaTypes: {
-            native: {
-              image: {
-                sizes: '300x250'
-              }
-            }
-          }
-        }];
-        result = checkAdUnitSetup(badNativeImgSize);
-        expect(result[0].mediaTypes.native.image.sizes).to.be.undefined;
-        expect(result[0].mediaTypes.native.image).to.exist;
-        sinon.assert.called(utils.logError);
-
-        let badNativeImgAspRat = [{
-          mediaTypes: {
-            native: {
-              image: {
-                aspect_ratios: '300x250'
-              }
-            }
-          }
-        }];
-        result = checkAdUnitSetup(badNativeImgAspRat);
-        expect(result[0].mediaTypes.native.image.aspect_ratios).to.be.undefined;
-        expect(result[0].mediaTypes.native.image).to.exist;
-        sinon.assert.called(utils.logError);
-
-        let badNativeIcon = [{
-          mediaTypes: {
-            native: {
-              icon: {
-                sizes: '300x250'
-              }
-            }
-          }
-        }];
-        result = checkAdUnitSetup(badNativeIcon);
-        expect(result[0].mediaTypes.native.icon.sizes).to.be.undefined;
-        expect(result[0].mediaTypes.native.icon).to.exist;
-        sinon.assert.called(utils.logError);
-      });
-    });
-  });
 });
